added passwords

+++ src/app/api/deadlines/route.ts
@@ -1,8 +1,16 @@
 import { NextResponse } from "next/server"
 import { sql } from "@vercel/postgres"
 import { revalidatePath, revalidateTag } from "next/cache"
+import { sha256 } from "js-sha256"
 
 export const dynamic = 'force-dynamic'
+
+function validatePassword(password: string){
+  console.log("validating")
+  //you though you could just look in the code and find the password ahaha bozo
+  const superSecretPassword = "1cf8531a22348eeb2abe9b891c939707aa5abfab759f1f6be7abc4a763968ebc"
+  return sha256(password) == superSecretPassword
+}
 
 export async function GET(request: Request){
   try{
@@ -20,8 +28,11 @@ export async function POST(request: Request){
     if (!request.body){
       return NextResponse.json({error: "Invalid request"}, {status: 500})
     }
-    let data: Deadline = await JSON.parse(await request.text())
+    let data: any = await JSON.parse(await request.text())
     console.log(data)
+    if (!validatePassword(data.password)){
+      return NextResponse.json({error: "Permission denied"}, {status: 401})
+    }
 
     const res = await sql`INSERT INTO deadlines
     (name, subject, start, due, mark, room, url, info)
@@ -41,6 +52,9 @@ export async function DELETE(request: Request){
   try{
     const data = await JSON.parse(await request.text())
     console.log("query")
+    if (!validatePassword(data.password)){
+      return NextResponse.json({error: "Permission denied"}, {status: 401})
+    }
     const res = await sql`DELETE FROM deadlines WHERE name = ${data.name} AND subject = ${data.subject}`
     revalidateTag("deadlines");
     console.log(data)
@@ -55,6 +69,9 @@ export async function DELETE(request: Request){
 export async function PUT(request: Request){
   try{
     const data = await JSON.parse(await request.text())
+    if (!validatePassword(data.password)){
+      return NextResponse.json({error: "Permission denied"}, {status: 401})
+    }
     const name = data.oldName
     const subject = data.oldSubject
